revise probabilistic PCA examples
